feat: Display calculated ticket details in output page
This commit adds a dedicated section to display the calculated ticket details (distance, age, discount type, discount amount, and final price) in the HTML page. The script now dynamically updates the content of the output page with these details after the "Calcola Prezzo" button is clicked. An initial message has been added to the output page to guide the user.